requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
aiogram==3.7.0
gspread==6.1.2
google-auth==2.30.0
qrcode[pil]==7.4.2
Pillow==10.3.0
httpx==0.27.0
pydantic==2.7.4
pydantic-settings==2.3.1
python-dotenv==1.0.1
user-agents==2.2.0
aiofiles==23.2.1
python-multipart==0.0.9
Jinja2==3.1.4
pytz==2024.1
